Modify Bootscene and GameScene classes
