+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/FrederickMih/Shooter-Game#readme",
   "devDependencies": {
     "babel-eslint": "^10.1.0",
+    "copy-webpack-plugin": "^9.0.0",
     "eslint": "^7.29.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.23.4",
