+++ src/Scenes/BootScene.js
@@ -1,0 +1,12 @@
+import Phaser from 'phaser';
+
+class BootScene extends Phaser.Scene {
+  constructor() {
+    super('Boot');
+  }
+
+  create() {
+    this.scene.start('Preloader');
+  }
+}
+export { BootScene as default };
